Use old alias logic for Rails 6.1.0.
Not thrilled that I’ve had to copy-paste code from pre-6.1.0 Rails, but it gets the tests passing, so it’ll do, at least for now. And it does have the advantage of decoupling a bit more joiner logic from potential future changes in Rails… so maybe it’s all for the best?